fix: 修复 post-install 脚本的错误

+++ post-install-hook.js
@@ -1,4 +1,7 @@
 const fs = require('fs');
+const { existsSync } = require('fs');
 
-fs.copyFileSync('./lib/wxmpAdaptation.js', './index.js');
-fs.unlink('./lib/wxmpAdaptation.js', () => {});
+if (existsSync('./lib/wxmpAdaptation.js')) {
+  fs.copyFileSync('./lib/wxmpAdaptation.js', './index.js');
+  fs.unlink('./lib/wxmpAdaptation.js', () => {});
+}
